CookieCutter/MLOps_Cookie/src/models/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
arrow=1.1.0=pyhd8ed1ab_1
binaryornot=0.4.4=py_1
blas=1.0=mkl
brotlipy=0.7.0=py38h294d835_1001
ca-certificates=2021.5.30=h5b45459_0
certifi=2021.5.30=py38haa244fe_0
cffi=1.14.5=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_1
click=8.0.1=py38haa244fe_0
colorama=0.4.4=pyh9f0ad1d_0
cookiecutter=1.7.3=pyh6c4a22f_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py_2
flake8=3.8.4=py_0
freetype=2.10.4=h546665d_1
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
intel-openmp=2021.2.0=haa95532_616
jbig=2.1=h8d14728_2003
jinja2=2.11.3=pyh44b312d_0
jinja2-time=0.2.0=py_2
jpeg=9d=h8ffe710_0
kiwisolver=1.3.1=py38hbd9d945_1
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.7=h8ffe710_5
libpng=1.6.37=h1d00b33_2
libtiff=4.3.0=h0c97f57_1
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h8ffe710_0
markupsafe=1.1.1=py38h294d835_3
matplotlib=3.4.2=py38haa244fe_0
matplotlib-base=3.4.2=py38heae8d8c_0
mccabe=0.6.1=py38_1
mkl=2021.2.0=haa95532_296
mkl-service=2.3.0=py38h2bbff1b_1
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.1=py38hf11a4ad_2
ninja=1.10.2=h6d14046_1
numpy=1.20.2=py38ha4e8547_0
numpy-base=1.20.2=py38hc2deb75_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb211442_1
openssl=1.1.1k=h8ffe710_0
pillow=8.2.0=py38h9273828_1
pip=21.1.1=py38haa95532_0
poyo=0.5.0=py_0
pycodestyle=2.6.0=py_0
pycparser=2.20=pyh9f0ad1d_2
pyflakes=2.2.0=py_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38haa244fe_7
pyqt-impl=5.12.3=py38h885f38d_7
pyqt5-sip=4.19.18=py38h885f38d_7
pyqtchart=5.12=py38h885f38d_7
pyqtwebengine=5.12.1=py38h885f38d_7
pysocks=1.7.1=py38haa244fe_3
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.1=py_0
python-slugify=5.0.2=pyhd8ed1ab_0
python_abi=3.8=1_cp38
pytorch=1.8.1=py3.8_cuda10.2_cudnn7_0
qt=5.12.9=h5909a2a_4
requests=2.25.1=pyhd3deb0d_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py_0
sqlite=3.35.4=h2bbff1b_0
text-unidecode=1.3=py_0
tk=8.6.10=h8ffe710_1
torchvision=0.2.2=py_3
tornado=6.1=py38h294d835_1
typing_extensions=3.7.4.3=pyha847dfd_0
unidecode=1.2.0=pyhd8ed1ab_0
urllib3=1.26.5=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
xz=5.2.5=h62dcd97_1
zipp=3.3.1=py_0
zlib=1.2.11=h62dcd97_1010
zstd=1.5.0=h6255e5f_0
